pyproject.toml
[tool.poetry]
name = "sqr-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
passlib = "1.7.4"
pyjwt = "2.8.0"
requests = "2.31.0"
streamlit = "1.33.0"
sqlalchemy = "1.4.52"
fastapi = "0.110.2"
uvicorn = "^0.29.0"
python-multipart = "^0.0.9"
pydantic-settings = "^2.2.1"
email-validator = "^2.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
